Iniciando o servico Restfull

+++ src/main/java/com/hudson/soares/clientes/clientes/model/ClienteDTO.java
@@ -6,6 +6,6 @@ import io.micrometer.common.lang.Nullable;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record ClienteDTO(@NotNull Integer id, @NotNull @NotBlank String nome, @NotBlank String cpf,
-        @Nullable @JsonIgnore Integer status_id, @Nullable String status_nome) {
+public record ClienteDTO(@NotNull Long id, @NotNull @NotBlank String nome, @NotBlank String cpf,
+        @Nullable @JsonIgnore Long status_id, @Nullable String status_nome) {
 }
